pyproject.toml
[tool.poetry]
name = "Bookmarks"
version = "1.0.0"
description = "Store bookmarks on Alfred"
documentation = "https://github.com/fedecalendino/alfred-bookmarks/blob/main/README.md"
homepage = "https://github.com/fedecalendino/alfred-bookmarks"
license = "MIT"

authors = [
    "Fede Calendino <fede@calendino.com>"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
alfred-pyflow = "0.4.6"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
ddt = "^1.6.0"
coverage = "^7.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
